+++ test/test_multipliers.py
@@ -1,7 +1,7 @@
 import pytest
 from openpile.construct import Pile, SoilProfile, Layer, Model
 from openpile.soilmodels import API_sand
-from openpile.analyze import winkler
+from openpile.winkler import winkler
 from openpile.utils.multipliers import durkhop
 
 import math as m